requirements.txt
streamlit==1.32.2
Pillow==10.1.0
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2